Added Punch & Stun Mechanics, Scoring Started
A new punch mechanic was added for the players + controller support for the new control.
Players can now be stunned when they are punched (This will disbale movement).
The scoring system has been started (Just UI text at the moment).
Fixed the collision with the finish line, so now the game ends when someone crosses it.
~Liam